+++ connect/src/index.ts
@@ -1,8 +1,8 @@
 export * from "./wormhole";
 export * from "./config";
-export * from "./wormholeTransfer";
 export * from "./common";
 
+export * from "./protocols/wormholeTransfer";
 export * from "./protocols/tokenTransfer";
 export * from "./protocols/cctpTransfer";
 export * from "./protocols/gatewayTransfer";
